# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is swagger for final project training golang mnc service with Hacktiv8",
        "title": "Final Project Training Golang MNC Service Api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Clara Velita Pranolo",
            "email": "claraavelitaa@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/users": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "This endpoint for delete user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.JSONSwaggerOKResponses"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "This endpoint for login user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "services",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.JSONSwaggerOKResponses"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "This endpoint for create new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register New User",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "services",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.JSONSwaggerCreatedResponses"
                        }
                    }
                }
            }
        },
        "/users/{userid}": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "This endpoint for update user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "payload",
                        "name": "services",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.JSONSwaggerOKResponses"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.JSONSwaggerCreatedResponses": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 201
                },
                "data": {},
                "errors": {},
                "status": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "dtos.JSONSwaggerOKResponses": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "errors": {},
                "status": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "dtos.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "dtos.UserRequest": {
            "type": "object",
            "required": [
                "age",
                "email",
                "password",
                "username"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dtos.UserUpdateRequest": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}